% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.AnalyzeData.R
\name{Analyze}
\alias{Analyze}
\title{Analyze all analyses using data collected from data centers.}
\usage{
Analyze(x)
}
\arguments{
\item{x}{\code{DistResNetReady} object prepared for analyses by \code{\link{RequestSiteDataPreparation}}.}
}
\value{
data frame populated with point estimates and variance estimates by various methods.
}
\description{
Extract regression data from each site, and conducts meta-analyses. Extracts summary-level data from each site, and conducts summary-level conditional regression. Extracts risk sets data from each site, and conducts case-centered conditional logistic regression. Extracts individual-level data without covariates
}
\author{
Kazuki Yoshida
}
