% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.AnalyzeData.R
\name{AnalyzeSiteDataset}
\alias{AnalyzeSiteDataset}
\title{Analyze individual-level data collected from data centers.}
\usage{
AnalyzeSiteDataset(x)
}
\arguments{
\item{x}{data frame generated by \code{\link{RequestSiteDataset}}, containing summary score matching, stratification, and weighting variables.}
}
\value{
data frame populated with point estimates and variance estimates by various methods.
}
\description{
Analyze individual-level data collected from data centers. Summary score matched analyses, stratified analyses, and weighted analyses are conducted stratifying on sites.
}
\author{
Kazuki Yoshida
}
