% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.AnalyzeData.R
\name{AnalyzeSiteDatasetBin}
\alias{AnalyzeSiteDatasetBin}
\title{Analyze individual-level binary data collected from data centers.}
\usage{
AnalyzeSiteDatasetBin(x)
}
\arguments{
\item{x}{data frame generated by \code{\link{RequestSiteDataset}}, containing summary score matching, stratification, and weighting variables.}
}
\value{
list populated with vectors containing a pair of a point estimate and variance estimate for each method.
}
\description{
Analyze individual-level binary data collected from data centers. PS- or DRS- matched analyses are conducted stratifying on sites. PS- or DRS-stratified analyses are conducted additionally stratifying on sites. PS-weighted analyses (stabilized IPTW and MW) are conducted stratifying on sites via \code{svycoxph} (conditional logisitc regression using a dummy constant time as the time variable).
}
\author{
Kazuki Yoshida
}
