% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.AnalyzeData.R
\name{AnalyzeSiteRegression}
\alias{AnalyzeSiteRegression}
\title{Meta-analyze regression results collected from data centers.}
\usage{
AnalyzeSiteRegression(x)
}
\arguments{
\item{x}{data frame generated by \code{\link{RequestSiteRegression}}}
}
\value{
data frame populated with point estimates and variance estimates by various methods. Columns for site-specific coefficients are also retained for detailed assessments.
}
\description{
Meta-analyze regression results collected from data centers. Pooling is conducted with inverse variance weighting (fixed effect meta-analysis).
}
\author{
Kazuki Yoshida
}
