% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.AnalyzeData.R
\name{AnalyzeSiteRisksets}
\alias{AnalyzeSiteRisksets}
\title{Analyze risk set-level data collected from data centers.}
\usage{
AnalyzeSiteRisksets(x)
}
\arguments{
\item{x}{data frame generated by \code{\link{RequestSiteRisksets}}}
}
\value{
data frame populated with point estimates and variance estimates by various methods.
}
\description{
Analyze risk set-level data collected from data centers. Case-centered conditional logistic regression is conducted conditioning on risk sets. Risk sets are finer than strata and sites, thus, strata and sites are effectively conditioned on.
}
\author{
Kazuki Yoshida
}
