% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.AnalyzeData.R
\name{AnalyzeSiteSummary}
\alias{AnalyzeSiteSummary}
\title{Analyze summary-level data collected from data centers.}
\usage{
AnalyzeSiteSummary(x)
}
\arguments{
\item{x}{data frame generated by \code{\link{RequestSiteSummary}}, containing summary-level data.}
}
\value{
data frame populated with point estimates and variance estimates by various methods.
}
\description{
Analyze summary-level data collected from data centers. Summary score matched analyses and stratified analyses.
}
\author{
Kazuki Yoshida
}
