% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.AnalyzeData.R
\name{AnalyzeSiteSummarySurv}
\alias{AnalyzeSiteSummarySurv}
\title{Analyze summary-level survival data collected from data centers.}
\usage{
AnalyzeSiteSummarySurv(x)
}
\arguments{
\item{x}{data frame generated by \code{\link{RequestSiteSummary}}, containing summary-level data.}
}
\value{
list populated with vectors containing a pair of a point estimate and variance estimate for each method.
}
\description{
Analyze summary-level survival data collected from data centers. PS- or DRS- matched analyses are conducted stratifying on sites. PS- or DRS-stratified analyses are conducted additionally stratifying on sites.
}
\author{
Kazuki Yoshida
}
