% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.AnalyzeData.R
\name{AnalyzeSiteTruth}
\alias{AnalyzeSiteTruth}
\title{Analyze individual-level data collected from data centers (counterfactuals)}
\usage{
AnalyzeSiteTruth(x, Tmax = 1)
}
\arguments{
\item{x}{data frame generated by \code{\link{RequestSiteTruth}}, containing all variable including latent ones (e.g., counterfactuals).}

\item{Tmax}{scalar value defining the end of follow up. This must match the one used in the observed data generation to calculate the relevant marginal hazard ratios.}
}
\value{
data frame populated with point estimates and variance estimates by various methods.
}
\description{
Analyze individual-level data collected from data centers. Summary score matched analyses, stratified analyses, and weighted analyses are conducted stratifying on sites.
}
\author{
Kazuki Yoshida
}
