% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{AssignCovariatesNormBin}
\alias{AssignCovariatesNormBin}
\title{Create indepedent covariates given mean, sd, and prevalence}
\usage{
AssignCovariatesNormBin(n, mean1, sd1, pp)
}
\arguments{
\item{n}{number of individuals to create}

\item{mean1}{mean for first continuous normal covariate}

\item{sd1}{sd for first continuous normal covariate}

\item{pp}{vector of prevalences for other Bernoulli covariates}
}
\value{
n x (m+1) data frame including ID and m covariates
}
\description{
The first covariate is a Normal(mean1, sd1) random variable, whereas the other
covariates are Bernoulli(p) random variables where p's are specified by the
elements of the pp vector.
}
\author{
Kazuki Yoshida
}
