% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{AssignOutcomeBin}
\alias{AssignOutcomeBin}
\title{Assign binary outcome based on covariates and treatment}
\usage{
AssignOutcomeBin(dfXA, beta0, betaX, betaA, betaXA)
}
\arguments{
\item{dfXA}{data frame including covariates and treatment assingment A}

\item{beta0}{scalar value coeffcient for the intercept}

\item{betaX}{vector specifying true coefficients for covariates}

\item{betaA}{scalar value for treatment effect}

\item{betaXA}{vector specifying true coefficients for covariates-treatment interaction. The vector must be the same length as \code{betaX}.}
}
\value{
data frame with additional counterfactual and observed variables related to the outcome.
}
\description{
A binary outcome is assigned as a Bernoulli(p_i) random variable where
p_i is an individual-specific outcome probability calculated from covariate
and treatment values and coefficients.
}
\author{
Kazuki Yoshida
}
