% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{AssignOutcomeSurv}
\alias{AssignOutcomeSurv}
\title{Assign survival time outcome}
\usage{
AssignOutcomeSurv(dfXA, betaX, betaA, betaXA, lambda, lambda_c, Tmax)
}
\arguments{
\item{dfXA}{data frame including covariates and treatment assingment A}

\item{betaX}{vector specifying true coefficients for covariates}

\item{betaA}{scalar value for treatment effect}

\item{betaXA}{vector specifying true coefficients for covariates-treatment interaction. The vector must be the same length as \code{betaX}.}

\item{lambda}{scalar value defining the constant baseline hazard for the event outcome. The corresponding time scale is in years.}

\item{lambda_c}{scalar value defining the constant baseline hazard for censoring. The corresponding time scale is in years.}

\item{Tmax}{scalar value defining the end of follow up. Use to introduce a constant value administrative censoring time. The corresponding time scale is in years. To avoid administrative censoring, set this to a very high value.}
}
\value{
data frame with added individual-specific rate (\code{rate}), counterfactual rates (\code{rate0} and \code{rate1}), event time (\code{T}; years), censoring time (\code{C}; years), observed time (\code{time}; integer days), and event indicator (\code{event}).
}
\description{
A time to event outcome is assigned as an Exponential(lambda_i) random
variable where lambda_i is an individual-specific time-constant hazard
of an event calculated from covariate and treatment values and coefficients.
Censoring times are generated from an Exponential(lambda_c_i) distribution
where lambda_c_i is an individual-specific time-constant hazard of censoring,
depending on the covariates only. The censoring time distribution does not
depend on the treatment to avoid collider stratification bias. An additional
administrative censoring time is defined by Tmax, which marks the end of the
follow up for those who have follow up times longer than Tmax.
The latent true event time and censoring time variables are coded in years.
The final observed time variable is coded in days (\code{ceiling} is applied
to avoid fractional numbers) to emulate the granularity of typical claims data.
}
\author{
Kazuki Yoshida
}
