% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{AssignTreatment}
\alias{AssignTreatment}
\title{Assign treatment based on true propensity score model}
\usage{
AssignTreatment(dfX, alpha0, alphaX)
}
\arguments{
\item{dfX}{data frame including covariates}

\item{alpha0}{a scalar value coeffcient for the intercept}

\item{alphaX}{a vector of length 7 specifiying true coefficients for 7 covariates}
}
\value{
data frame with additional true propensity score pA and treatment A.
}
\description{
A binary treatment is assigned as a Bernoulli(p_i) random variable where
p_i is an individual-specific treatment probability (true propensity score)
calculated from covariate values and coefficients.
}
\author{
Kazuki Yoshida
}
