% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.AnalyzeData.R
\name{CaseCenteredLogistic}
\alias{CaseCenteredLogistic}
\title{Conduct case-centered analysis}
\usage{
CaseCenteredLogistic(df)
}
\arguments{
\item{df}{data frame containing one row for each riskset with the following variables: \code{events_A1} (event count among the exposed), \code{events_A0} (event count among the unexposed), \code{riskset_A1} (total count among the exposed including those who are having events), and \code{riskset_A0} (total count among the unexposed including those who are having events).}
}
\value{
\code{glm} fit object if successfully fit.
}
\description{
Given a specifically formatted data frame, conduct case-centered analysis
}
\author{
Kazuki Yoshida
}
