% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.AnalyzeData.R
\name{CaseCenteredLogisticWeightedBreslow}
\alias{CaseCenteredLogisticWeightedBreslow}
\title{Conduct case-centered analysis for weighted risk sets (Breslow)}
\usage{
CaseCenteredLogisticWeightedBreslow(df)
}
\arguments{
\item{df}{data frame containing one row for each riskset with the following variables: \code{events_A1} (event count among the exposed), \code{events_A0} (event count among the unexposed), \code{riskset_A1} (total count among the exposed including those who are having events), and \code{riskset_A0} (total count among the unexposed including those who are having events). The data frame additionally must contain valid weighted versions of these variables (\code{w_events_A1}, etc) and variance of weights for four types of cells defined by event and exposure status (\code{varw_events_A*} and \code{varw_nonevents_A*}), approximate weighted analysis is conducted.}
}
\value{
\code{svycoxph} fit object if successfully fit.
}
\description{
Given a specifically formatted weighted data frame, conduct case-centered analysis by weighted individual-level data reconstruction. This one uses Breslow approximation.
}
\author{
Kazuki Yoshida
}
