% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{ConstructAssignCovariatesNormBin}
\alias{ConstructAssignCovariatesNormBin}
\title{Construct independent covariate generator}
\usage{
ConstructAssignCovariatesNormBin(mean1, sd1, p2, pLast)
}
\arguments{
\item{mean1}{mean for first continuous normal covariate}

\item{sd1}{sd for first continuous normal covariate}

\item{p2}{prevalence of binary X2}

\item{pLast}{prevalence of binary Xlast}
}
\value{
covariate generator that takes n and p as arguments
}
\description{
Construct an independent covariate generator that generates the first covariate as Normal(mean1,sd1) random variable, whereas the other covariates are Bernoulli(p) random variables where p's are specified by equal interval sequence of (p2, ..., pp)
}
\author{
Kazuki Yoshida
}
