% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{CreateIpw}
\alias{CreateIpw}
\title{Create IPW}
\usage{
CreateIpw(score, status, type)
}
\arguments{
\item{score}{vector of predicted probability of being assigned \code{status == 1}.}

\item{status}{vector of observed status (binary 0, 1).}

\item{type}{character value containing either one of \code{naive} (non-stabilized naive IPW), \code{stabilized} (IPW stabilized by marginal probabilities of \code{status}), or \code{mw} (matching weights; stabilized by \code{pmin(score, 1 - score)}).}
}
\value{
vector of IPW.
}
\description{
Create inverse probability weights given probability of status assignment and observed status.
}
\author{
Kazuki Yoshida
}
