% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{CreateMatches}
\alias{CreateMatches}
\title{Match across status}
\usage{
CreateMatches(score, status, logit, caliper)
}
\arguments{
\item{score}{summary score vector used as matching metric.}

\item{status}{binary status variable vector to match across.}

\item{logit}{whether to perform matching on the logit scale. Use if using scores spanning [0,1] only.}

\item{caliper}{caliper multipliation factor. The actual caliper is \code{caliper * sd(score)}.}
}
\value{
vector of matched pair IDs containing NA for unmatched.
}
\description{
Match across status
}
\author{
Kazuki Yoshida
}
