% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.Miscellaneous.R
\name{CreateNaDf}
\alias{CreateNaDf}
\title{Generate a desired NA data frame}
\usage{
CreateNaDf(labels, as_funcs, nrow)
}
\arguments{
\item{labels}{character vector for column names}

\item{as_funcs}{vector transforming functions such as \code{as.integer}, etc}

\item{nrow}{scalar value to specify the number of rows}
}
\value{
a data frame filled with |code{NA}'s and \code{nrow} rows with column names as specified in \code{labels} and column classes as specified in \code{as_funcs}.
}
\description{
Generate a desired NA data frame having specified number of rows, column names, and column classes.
}
\author{
Kazuki Yoshida
}
