% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{CreateStrata}
\alias{CreateStrata}
\title{Create strata}
\usage{
CreateStrata(score, nStrata)
}
\arguments{
\item{score}{summary score to be categorized.}

\item{nStrata}{number of strata.}
}
\value{
vector strata as numeric
}
\description{
Create strata by categorizing summary scores into \code{nStrata} quantiles. If there are fewer discrete scores than \code{nStrata}, the rank will be returned. Sometimes a single value may have many data points causing quantile threshold to be identical values. The duplicated threshold values are dropped in this case, giving levels fewer than nStrata.
}
\author{
Kazuki Yoshida
}
