% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{distributed-package}
\alias{distributed-package}
\alias{distributed}
\title{Simulation suite for privacy preserving data analysis methods}
\description{
Implement simulation for privacy preserving data analysis methods in the setting of distributed data network.
}
\author{
Kazuki Yoshida
}
