% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{EstimateScoreBin}
\alias{EstimateScoreBin}
\title{Estimate summary confounder scores (binary logistic model)}
\usage{
EstimateScoreBin(formula, data, newdata)
}
\arguments{
\item{formula}{formula to specify the model used to estimate the summary score. For propensity score, specify the treatment model. For disease risk score, specify the outcome model in the absence of treatment.}

\item{data}{data frame of data for model building. Specify the entire dataset for propensity score estimation. Specify the untreated subset for disease risk score estimation.}

\item{newdata}{data frame of data for prediction. Specify the entire dataset.}
}
\value{
vector of summary scores predicted for \code{newdata}.
}
\description{
Estimate propensity score or disease risk score depending on the model specification. This is for binary outcomes.
}
\author{
Kazuki Yoshida
}
