% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{EstimateScoreSurv}
\alias{EstimateScoreSurv}
\title{Estimate disease risk score (Cox model)}
\usage{
EstimateScoreSurv(formula, data, newdata)
}
\arguments{
\item{formula}{formula to specify the model used to estimate the summary score. For disease risk score, specify the outcome model in the absence of treatment.}

\item{data}{data frame of data for model building. Specify the untreated subset for disease risk score estimation.}

\item{newdata}{data frame of data for prediction. Specify the entire dataset.}
}
\value{
vector of linear predictors (log hazard ratios) for \code{newdata}.
}
\description{
Estimate disease risk score using Cox proportional hazards model. This is for survival outcomes. The disease risk score is defined as the linear predictor from the \code{coxph} model object. This is the log hazard ratio (range [-Inf, Inf]) of the event compared to the mean covariate individual.
}
\author{
Kazuki Yoshida
}
