% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.Miscellaneous.R
\name{ExtremeTwoPointSample}
\alias{ExtremeTwoPointSample}
\title{Generate a two point distribution sample with only one point above mean.}
\usage{
ExtremeTwoPointSample(n, m, v)
}
\arguments{
\item{n}{Required number of individuals in the sample. If 0, an empty row data frame is returned.}

\item{m}{Required positive sample mean. As this is used to reconstruct a weight distribution, it has to be positive. NaN created by 0/0 is allowed if n = 0.}

\item{v}{Required non-negative sample variance. As the weights can be the same for all individuals, zero is permitted.}
}
\value{
data frame with a \code{W} column containing the values and a \code{count} column containing the number of subjects having value \code{W}.
}
\description{
Generate a two point distribution sample with only one point above mean with the desired mean and variance.
}
\author{
Kazuki Yoshida
}
