% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{GenerateDataForAllScenarios}
\alias{GenerateDataForAllScenarios}
\title{Generate R iterations for each scenario}
\usage{
GenerateDataForAllScenarios(Scenarios, parts, R)
}
\arguments{
\item{Scenarios}{a Scenarios object. Each list element is a specification for a given scenario.}

\item{parts}{how many subfiles to create for each scenario. Use this to ease parallelization across cluster nodes.}

\item{R}{iteration count for each subfile. The total iteration count for a given scenario is \code{parts * R}.}
}
\value{
Use for its side effect. There is no return value. Data files are created in the same folder.
}
\description{
Given an iteration count and a list of scenario parameters for multiple scenarios (\code{Scenarios} object), generate R datasets for each scenario. If a scenario within a \code{Scenarios} object is corrupt (not correctly a \code{ScenarioDistResNet} object), it is skipped by \code{try}.
}
\author{
Kazuki Yoshida
}
