% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{GenerateDistResNet}
\alias{GenerateDistResNet}
\title{Generate a distributed research network.}
\usage{
GenerateDistResNet(lstN, lstAssignCovariates, lstAlphas, lstBetas,
  lstSurvParams)
}
\arguments{
\item{lstN}{list of sizes of centers. Number of centers (K) is also determined by its length.}

\item{lstAssignCovariates}{list of covariate generation functions that takes n and p as the only arguments.}

\item{lstAlphas}{list of alpha vectors potentially varying among centers. See \code{\link{GenerateOneCenter}} for construction of an alpha vector.}

\item{lstBetas}{list of beta vectors potentially varying among centers. See \code{\link{GenerateOneCenter}} for construction of a beta vector.}

\item{lstSurvParams}{list of survival outcome parameters potentially varying among centers.}
}
\value{
list of K data frames, each of which contains a complete dataset from a single data center
}
\description{
Generate a distributed research network of K sites.
}
\author{
Kazuki Yoshida
}
