% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{GenerateOneCenter}
\alias{GenerateOneCenter}
\title{Generate one data center}
\usage{
GenerateOneCenter(n, alphas, betas, survParams)
}
\arguments{
\item{n}{study site-specific sample size}

\item{alphas}{parameter vector for treatment model including c(alpha0, alphaX)}

\item{betas}{parameter vector for outcome model shared among binary and survival outcome models including \code{c(beta0, betaX, betaA, betaXA)}.}

\item{survParams}{vector of two. The first element is the baseline hazard of events in the exponential event time outcome model (\code{lambda}). The second element is the baseline hazard of censoring in the exponential censoring time model (\code{lambda_c}).}
}
\value{
data frame with id, covariates, treatment (A) related variables, binary outcome (Y) related variables, and survival outcome (time, event) related variables.
}
\description{
Generate one data center with information of n patients. The parameters are given as four raw vectors. This function is intentionally designed to be a low level function working with raw numeric vector parameters for maximum flexibility. In actual simulation runs, this function should be called by \code{\link{GenerateDistResNet}}, not directly.
}
\author{
Kazuki Yoshida
}
