% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.AnalyzeData.R
\name{GenerateSasProcPhreg}
\alias{GenerateSasProcPhreg}
\title{Generate a SAS script to run PROC PHREG}
\usage{
GenerateSasProcPhreg(datafile, method, strata, outfile)
}
\arguments{
\item{datafile}{csv input data file path}

\item{method}{Either one of efron or breslow.}

\item{strata}{Specify space-deliminated names of stratifying variables.}

\item{outfile}{csv out data file path}
}
\value{
vector containing the name of SAS script file and planned output csv file name.
}
\description{
Generate a SAS script string with hard-coded input file, output file, approximation method, and stratification variables.
}
\author{
Kazuki Yoshida
}
