% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.Miscellaneous.R
\name{is.error}
\alias{is.error}
\title{Tests for objects of type \code{try-error}}
\usage{
is.error(x)
}
\arguments{
\item{x}{return object from a function wrapped by \code{try}.}
}
\value{
logical indicating whether the return object is an error object given by \code{try} or normal return object.
}
\description{
Tests for objects of type \code{try-error}
}
\author{
Kazuki Yoshida
}
