% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{print.DistResNet}
\alias{print.DistResNet}
\title{print method for a DistResNet class object}
\usage{
\method{print}{DistResNet}(x, ...)
}
\arguments{
\item{x}{DistResNet object}

\item{...}{for compatibility with generic functions.}
}
\value{
NULL. Print results
}
\description{
Run the print method for each center (ResSite) and print results.
}
\author{
Kazuki Yoshida
}
