% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{print.ParamsTableOne}
\alias{print.ParamsTableOne}
\title{print method for the ParamsTableOne object}
\usage{
\method{print}{ParamsTableOne}(x, print = TRUE, ...)
}
\arguments{
\item{x}{ParamsTableOne object generated by \code{\link{summary.ResSite}} or \code{\link{GenerateOneCenterTable}}.}

\item{print}{whether to print results}

\item{...}{additional parameters passed to \code{print.TableOne}}
}
\value{
return a matrix object. Print results.
}
\description{
Print the table summarizing variables. If included, parameter values are also printed. Invisibly return a matrix containing the table.
}
\author{
Kazuki Yoshida
}
