% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{print.ResSite}
\alias{print.ResSite}
\title{print method for a ResSite class object}
\usage{
\method{print}{ResSite}(x, ...)
}
\arguments{
\item{x}{ResSite object to print}

\item{...}{for compatibility with generic functions.}
}
\description{
Print the sample size and analysis readiness of a research site in a distributed research network.
}
\author{
Kazuki Yoshida
}
