% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{print.ScenarioDistResNet}
\alias{print.ScenarioDistResNet}
\title{print method for a ScenarioDistResNet class object}
\usage{
\method{print}{ScenarioDistResNet}(x, ...)
}
\arguments{
\item{x}{ScenarioDistResNet object}

\item{...}{for compatibility with generic functions.}
}
\value{
NULL. Print results
}
\description{
Print parameters for a scenario for a distributed research network.
}
\author{
Kazuki Yoshida
}
