% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{print.Scenarios}
\alias{print.Scenarios}
\title{print method for a Scenarios class object}
\usage{
\method{print}{Scenarios}(x, ...)
}
\arguments{
\item{x}{Scenarios object}

\item{...}{for compatibility with generic functions.}
}
\value{
NULL. Print results
}
\description{
Print parameters for each scenario included in a set of scenarios (Scenarios object).
}
\author{
Kazuki Yoshida
}
