% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.AggregateAcrossSites.R
\name{RequestSiteDataPreparation}
\alias{RequestSiteDataPreparation}
\title{Request sites perform within-site data preparation.}
\usage{
RequestSiteDataPreparation(x)
}
\arguments{
\item{x}{\code{DistResNet} object representing a distributed research network.}
}
\value{
\code{DistResNetReady} object that has been prepared for other \code{RequestSite*} functions.
}
\description{
Invokes \code{\link{SitePrepareHelperVariables}} on each \code{ResSite} object within a \code{DistResNet} object. Data preparation is performed within each research site so that further analyses can be conducted. This process includes summary score estimation, matching, stratification, and weighting. other \code{RequestSite*} functions can then ask for sharing of relevant data across sites.
}
\author{
Kazuki Yoshida
}
