% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.AggregateAcrossSites.R
\name{RequestSiteRisksets}
\alias{RequestSiteRisksets}
\title{Request sites provide site-specific risk set data.}
\usage{
RequestSiteRisksets(x)
}
\arguments{
\item{x}{\code{DistResNetReady} object representing a distributed research network.}
}
\value{
data frame containing results for the binary outcome and survival outcome.
}
\description{
Request sites provide site-specific risk set data. Aggregated risk set data from all the participating sites are returned.
}
\author{
Kazuki Yoshida
}
