% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.AggregateAcrossSites.R
\name{RequestSiteTruth}
\alias{RequestSiteTruth}
\title{Request sites provide site-specific individual-level data (counterfactuals).}
\usage{
RequestSiteTruth(x)
}
\arguments{
\item{x}{\code{DistResNetReady} object representing a distributed research network.}
}
\value{
data frame containing individual-level observed data.
}
\description{
Request sites provide site-specific individual-level data. Aggregated individual-level data from all the participating sites are returned. This version will request the latent variables including counterfactuals.
}
\author{
Kazuki Yoshida
}
