% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.AnalyzeData.R
\name{RisksetsToIpd}
\alias{RisksetsToIpd}
\title{Expand weighted risk set data with variance to long format.}
\usage{
RisksetsToIpd(x, compress = FALSE)
}
\arguments{
\item{x}{data frame generated by \code{\link{RequestSiteRisksets}}, containing risk set-level data.}

\item{compress}{defaults to \code{FALSE}. If true, the summary long-format data containing one row for each unique combination of variables including weights are created. The count variable indicates how many individuals each row represent.}
}
\value{
data frame populated with point estimates and variance estimates by various methods.
}
\description{
Expand weighted risk set-level data to individual-level data or summary long-format data. The weights are regenerated to maintain the given mean and variance. The individul-level data contain one row for each individual in each risk set. The summary long-format data are compressed so that each row represent multiple individuals. The count variable indicates how many individuals each row represent.
}
\author{
Kazuki Yoshida
}
