% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.AnalyzeData.R
\name{RisksetsToIpdSkewed}
\alias{RisksetsToIpdSkewed}
\title{Expand weighted risk set data with variance to long format (enhanced).}
\usage{
RisksetsToIpdSkewed(x, compress, helper_fun = RisksetsToIpdSkewedHelper)
}
\arguments{
\item{x}{data frame generated by \code{\link{RequestSiteRisksets}}, containing risk set-level data.}

\item{compress}{defaults to \code{FALSE}. If true, the summary long-format data containing one row for each unique combination of variables including weights are created. The count variable indicates how many individuals each row represent.}

\item{helper_fun}{helper function used to regenerate weights given sample size \code{n}, sample mean \code{m}, sample variance \code{v}, and labeling for treatment \code{A} and event status \code{event}. Defaults to \code{RisksetsToIpdSkewedHelper}, which tries to balance the number of observations with above-mean and below-mean weights. \code{RisksetsToIpdExtremeHelper} will positions just one observation with an above-mean weight and all others below mean.}
}
\value{
data frame populated with point estimates and variance estimates by various methods.
}
\description{
Expand weighted risk set-level data to individual-level data or summary long-format data. The weights are regenerated to maintain the given mean and variance. This version specifically try to avoid negative weights by skewing distribution. Also computation is based on the summary version, which is expaned when \code{compress = FALSE}. The individul-level data contain one row for each individual in each risk set. The summary long-format data are compressed so that each row represent multiple individuals. The count variable indicates how many individuals each row represent.
}
\author{
Kazuki Yoshida
}
