% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.SummarizeResults.R
\name{SeqGsub}
\alias{SeqGsub}
\title{Sequentially replace strings}
\usage{
SeqGsub(x, pat_rep_vec)
}
\arguments{
\item{x}{vector of strings. Transformed to character if of other type.}

\item{pat_rep_vec}{vector with even number of elements. Each odd number pattern is replaced with the following even number string. The replacement proceed sequentially, so the order of these pattern-replacement pairs matter.}
}
\value{
a vector with replaced strings
}
\description{
Wrapper for multiple chained \code{gsub} replacement. See the explanation of \code{pat_rep_vec} for details
}
\author{
Kazuki Yoshida
}
