% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{SiteDataset}
\alias{SiteDataset}
\title{Run site-specific individual-level data extraction}
\usage{
SiteDataset(df)
}
\arguments{
\item{df}{data frame for a single data center preapred by \code{\link{SitePrepareHelperVariables}}. It must have class \code{ResSiteReady}.}
}
\value{
data frame containing individual-level observed data.
}
\description{
Run site-specific individual-level data extraction and return a data frame.
}
\author{
Kazuki Yoshida
}
