% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{SitePrepareHelperVariables}
\alias{SitePrepareHelperVariables}
\title{Prepare PS- and DRS-related helper variables}
\usage{
SitePrepareHelperVariables(x)
}
\arguments{
\item{x}{ResSite object containing a dataset from a single center given by \code{\link{GenerateOneCenter}}. All covariates (variable names starting with X) are used in PS and DRS estimation.}
}
\value{
ResSiteReady object, a data frame with additional PS, DRS, strata, and match indicator.
}
\description{
Add estimated PS and DRS to the dataset as well as a match indicator, weights, and strata based on these summary scores.
}
\author{
Kazuki Yoshida
}
