% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{SiteRegressionBin}
\alias{SiteRegressionBin}
\title{Run site-specific logistic regression}
\usage{
SiteRegressionBin(df)
}
\arguments{
\item{df}{data frame for a single data center preapred by \code{\link{SitePrepareHelperVariables}}. It must have class \code{ResSiteReady}.}
}
\value{
data frame containing point estimates and variance estimates.
}
\description{
Run site-specific logistic regression using both PS and DRS. Matched analyses, stratified analyses (conditional logistic regression via \code{survival::clogit}), and weighted analyses (sandwich variance estimator via the \code{survey} package) are conducted.
}
\author{
Kazuki Yoshida
}
