% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{SiteRegressionSurv}
\alias{SiteRegressionSurv}
\title{Run site-specific Cox regression}
\usage{
SiteRegressionSurv(df)
}
\arguments{
\item{df}{data frame for a single data center preapred by \code{\link{SitePrepareHelperVariables}}. It must have class \code{ResSiteReady}.}
}
\value{
list of pairs of a point estimate and a variance estimate for the treatment effect.
}
\description{
Run site-specific Cox regression using both PS and DRS. Matched analyses, stratified analyses, and weighted analyses are conducted.
}
\author{
Kazuki Yoshida
}
