% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{SiteRisksets}
\alias{SiteRisksets}
\title{Run site-specific risk set data creation}
\usage{
SiteRisksets(df)
}
\arguments{
\item{df}{data frame for a single data center preapred by \code{\link{SitePrepareHelperVariables}}. It must have class \code{ResSiteReady}.}
}
\value{
data frame containing results for the binary and survival outcomes.
}
\description{
Run site-specific risk set data creation and return a data frame.
}
\author{
Kazuki Yoshida
}
