% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{SiteRisksetsBin}
\alias{SiteRisksetsBin}
\title{Obtain site-specific risk sets for binary outcome}
\usage{
SiteRisksetsBin(df)
}
\arguments{
\item{df}{data frame for a single data center preapred by \code{\link{SitePrepareHelperVariables}}. It must have class \code{ResSiteReady}.}
}
\value{
list containing results for PS/DRS-matched risk sets and PS/DRS-stratified risk sets.
}
\description{
Obtain site-specific risk sets for binary outcome using PS and DRS-matched cohorts as well as PS and DRS-stratified cohorts.
}
\author{
Kazuki Yoshida
}
