% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{SiteRisksetsSurv}
\alias{SiteRisksetsSurv}
\title{Obtain site-specific risk sets for survival outcome}
\usage{
SiteRisksetsSurv(df)
}
\arguments{
\item{df}{data frame for a single data center preapred by \code{\link{SitePrepareHelperVariables}}. It must have class \code{ResSiteReady}.}
}
\value{
data frame containing point estimates and variance estimates.
}
\description{
Obtain site-specific risk sets for survival outcome using PS and DRS-matched cohorts as well as PS and DRS-stratified cohorts.
}
\author{
Kazuki Yoshida
}
