% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{SiteSummarySurv}
\alias{SiteSummarySurv}
\title{Obtain site-specific summaries for survival outcome}
\usage{
SiteSummarySurv(df)
}
\arguments{
\item{df}{data frame for a single data center preapred by \code{\link{SitePrepareHelperVariables}}. It must have class \code{ResSiteReady}.}
}
\value{
data frame containing results for PS/DRS-matched cohorts and PS/DRS-stratified cohorts.
}
\description{
Obtain site-specific summaries for survival outcome using PS and DRS-matched cohorts as well as PS and DRS-stratified cohorts.
}
\author{
Kazuki Yoshida
}
