% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{SiteTruth}
\alias{SiteTruth}
\title{Run site-specific individual-level data extraction (counterfactuals)}
\usage{
SiteTruth(df)
}
\arguments{
\item{df}{data frame for a single data center preapred by \code{\link{SitePrepareHelperVariables}}. It must have class \code{ResSiteReady}.}
}
\value{
data frame containing individual-level data including counterfactuals.
}
\description{
Run site-specific individual-level data extraction including all variables both observed and latent (e.g., counterfactuals).
}
\author{
Kazuki Yoshida
}
