% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{SummarizeEventsPtBy}
\alias{SummarizeEventsPtBy}
\title{Summarize event counts and total person-time by groups.}
\usage{
SummarizeEventsPtBy(event, time, by, data)
}
\arguments{
\item{event}{variable name of a binary event indicator variable (coded 0 or 1) in the data frame.}

\item{time}{variable name of a follow-up time variable in the data frame.}

\item{by}{variable name of a categorical variable used as a grouping variable. This should usually be a factor if called in a stratified analysis to avoid differential lengths.}

\item{data}{data frame containing these variables.}
}
\value{
data frame containing the \code{events} (event count) element the \code{denom} (person-time) information for each treatment group. When the data frame supplied is empty (0 rows), a single-row data frame filled with NA is returned.
}
\description{
Summarize event counts and total person-time of survival outcome (coded 0 or 1) classified
by a grouping variable. Intended for summarizing these counts by the binary
treatment groups. When there are empty cells, the event counts and the total
person-time are both forced to be zeros.
}
\author{
Kazuki Yoshida
}
