% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{SummarizeEventsTotalsBy}
\alias{SummarizeEventsTotalsBy}
\title{Summarize event counts and trial counts by groups.}
\usage{
SummarizeEventsTotalsBy(var, by, data)
}
\arguments{
\item{var}{variable name of a binary variable (coded 0 or 1) in the data frame.}

\item{by}{variable name of a categorical variable used as a grouping variable. This should usually be a factor if called in a stratified analysis to avoid differential lengths.}

\item{data}{data frame containing these variables.}
}
\value{
data frame containing the \code{events} (event count) and the \code{denom} (trial count) information for each treatment group. When the data frame supplied is empty (0 rows), a single-row data frame filled with NA is returned.
}
\description{
Summarize event counts and trial counts of binary outcome (coded 0 or 1) classified
by a grouping variable. Intended for summarizing these counts by the binary
treatment groups. When there are empty cells, the event counts and the total
counts are both forced to be zeros.
}
\author{
Kazuki Yoshida
}
