% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.ParepareDataWithinSites.R
\name{SummarizeStratumEventsTotalsBy}
\alias{SummarizeStratumEventsTotalsBy}
\title{Summarize stratum-specific event counts and trial counts by groups.}
\usage{
SummarizeStratumEventsTotalsBy(var, by, strata, data)
}
\arguments{
\item{var}{variable name of a binary variable (coded 0 or 1) in the data frame.}

\item{by}{variable name of a categorical variable used as a grouping variable. This should usually be a factor if called in a stratified analysis to avoid differential lengths.}

\item{strata}{variable name of a stratifying variable.}

\item{data}{data frame containing these variables.}
}
\value{
data frame containing one stratum and treatment-specific row of \code{events} (event count) and \code{denom} (trial count) for each stratum. When the data frame supplied is empty (0 rows), a single-row data frame filled with NA is returned.
}
\description{
Summarize stratum-specific event counts and trial counts by groups and return
a list of lists containing events and totals.
}
\author{
Kazuki Yoshida
}
