% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{summary.DistResNet}
\alias{summary.DistResNet}
\title{summary method for a DistResNet class object}
\usage{
\method{summary}{DistResNet}(x, truth = FALSE, ...)
}
\arguments{
\item{x}{DistResNet object}

\item{truth}{whether to include the unobserved latent variables including counterfactual probabilities of the binary outcome assignment, counterfactual rates of the survival events, as well as true underlying 1-year survival \code{S1yr}.}

\item{...}{for compatibility with generic functions.}
}
\value{
a list of \code{ParamsTableOne} objects.
}
\description{
Summarizes variables for each research site in a distributed research network.
}
\author{
Kazuki Yoshida
}
