% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.GenerateData.R
\name{summary.ResSite}
\alias{summary.ResSite}
\title{summary method for a ResSite class object}
\usage{
\method{summary}{ResSite}(x, truth = FALSE, ...)
}
\arguments{
\item{x}{ResSite object to print}

\item{truth}{whether to include the unobserved latent variables including counterfactual probabilities of the binary outcome assignment, counterfactual rates of the survival events, as well as true underlying 1-year survival \code{S1yr}.}

\item{...}{for compatibility with generic functions.}
}
\value{
\code{ParamsTableOne} object containing parameter values (if \code{truth = TRUE}) and TableOne object defined in \code{tableone}. If \code{truth} is TRUE, then it will include parameters, latent variables, and the true one-year survival probability.
}
\description{
Summarizes variables in a research site in a distributed research network.
}
\author{
Kazuki Yoshida
}
