% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.Miscellaneous.R
\name{try_w}
\alias{try_w}
\title{Enhanced try that can treat some warnings as errors}
\usage{
try_w(expr, regexp = NULL, silent = FALSE)
}
\arguments{
\item{expr}{expression to be evaluated that may give an warning or error.}

\item{regexp}{regular expression to specify warning to be handled as an error.}

\item{silent}{whether to silence error/warning messages.}
}
\value{
\code{try-error} object if an error or a specified warning is encountered. Otherwise, the result of evaluation returns.
}
\description{
Behaves like \code{try} to return \code{try-error} object on an error. If a warning is encountered it is checked against the specified regular expression \code{regexp}. If the warning message matches the regular expression, it is handled as an error and return \code{try-error} object.
}
\author{
Kazuki Yoshida
}
