% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.Miscellaneous.R
\name{ValidateDf}
\alias{ValidateDf}
\title{Validate data frame for variability in exposure and outcome}
\usage{
ValidateDf(df, expo_var, bin_var = NULL, event_var = NULL)
}
\arguments{
\item{df}{data frame to be used in a regression analysis}

\item{expo_var}{string for the exposure variable name}

\item{bin_var}{string for the binary outcome variable name}

\item{event_var}{string for the survival event indicator variable name (1 for event)}
}
\value{
the validated data frame if it is valid. Otherwise, an error is raised to stop estimation on invalid data.
}
\description{
Validate a given data frame for variability in the exposure variable and outcome variable. The exposure variable specified in \code{expo_var} should have two values to make regression meaningful. Only one of \code{out_var} and \code{event_var} is required although both can be validated at the same time. Because the error raised can stop the simulation, it should be used within \code{try} function.
}
\author{
Kazuki Yoshida
}
